🐛 Se ha cambiado el id de bigint a long en RoomImage.

+++ src/main/java/com/example/skystayback/models/RoomImage.java
@@ -12,7 +12,7 @@ public class RoomImage {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Column(name = "id", nullable = false, unique = true)
-    private BigInteger id;
+    private Long id;
 
     @ManyToOne(cascade = CascadeType.PERSIST, fetch = FetchType.EAGER)
     @JoinColumn(name = "room_id")
